+++ src/modules/ui/IMGUIApp.h
@@ -118,7 +118,7 @@ protected:
 	virtual bool onKeyPress(int32_t key, int16_t modifier) override;
 	virtual bool onTextInput(const core::String& text) override;
 	virtual void onMouseMotion(void *windowHandle, int32_t x, int32_t y, int32_t relX, int32_t relY) override;
-	virtual bool onMouseWheel(int32_t x, int32_t y) override;
+	virtual bool onMouseWheel(float x, float y) override;
 	virtual void onMouseButtonRelease(int32_t x, int32_t y, uint8_t button) override;
 	virtual void onMouseButtonPress(int32_t x, int32_t y, uint8_t button, uint8_t clicks) override;
 	virtual bool handleSDLEvent(SDL_Event& event) override;
