+++ src/modules/video/EventHandler.h
@@ -61,7 +61,7 @@ public:
 	 * @param[in] id game controller id
 	 */
 	void controllerMotion(uint8_t axis, int value, uint32_t id);
-	void mouseWheel(int32_t x, int32_t y);
+	void mouseWheel(float x, float y);
 	void mouseMotion(void *windowHandle, int32_t x, int32_t y, int32_t relX, int32_t relY);
 	void mouseButtonPress(int32_t x, int32_t y, uint8_t button, uint8_t clicks);
 	void mouseButtonRelease(int32_t x, int32_t y, uint8_t button);
