+++ src/modules/video/WindowedApp.cpp
@@ -214,15 +214,15 @@ bool WindowedApp::handleKeyPress(int32_t key, int16_t modifier, uint16_t count) 
 	return _keybindingHandler.execute(key, modifier, true, nowSeconds(), count);
 }
 
-bool WindowedApp::onMouseWheel(int32_t x, int32_t y) {
+bool WindowedApp::onMouseWheel(float x, float y) {
 	int32_t key;
-	if (y < 0)
+	if (y < 0.0f)
 		key = util::button::CUSTOM_SDLK_MOUSE_WHEEL_UP;
-	else if (y > 0)
+	else if (y > 0.0f)
 		key = util::button::CUSTOM_SDLK_MOUSE_WHEEL_DOWN;
-	else if (x < 0)
+	else if (x < 0.0f)
 		key = util::button::CUSTOM_SDLK_MOUSE_WHEEL_LEFT;
-	else if (x > 0)
+	else if (x > 0.0f)
 		key = util::button::CUSTOM_SDLK_MOUSE_WHEEL_RIGHT;
 	else
 		return false;
